finance-llm-service/requirements.txt
flask
ai21
python-dotenv